working on user registrations

+++ base/views.py
@@ -1,7 +1,14 @@
 from django.shortcuts import render,redirect
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q #for searching purpose
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room,Topic
 from .forms import RoomForm
+
+from django.contrib.auth import authenticate,login,logout
 # Create your views here.
 
 # rooms =  [
@@ -9,6 +16,55 @@ from .forms import RoomForm
 #     {'id':2,'name': 'Design with me'},
 #     {'id':3,'name': 'Frontend developers'},
 # ]
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:# if the user tries to login using url while being logged in restrict them
+        return redirect('home')
+    
+    if request.method =="POST": # checking if user has entered the infos and logged in 
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try: # checking if the user already exist
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User does not exist')
+
+        user = authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+              messages.error(request,'Username or password does not exist')
+
+
+
+    context={'page':page}
+
+    return render(request,'base/login_register.html',context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method =='POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower() # making sure the username is lowercase
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occurred during registration')
+
+    return render(request,'base/login_register.html',{'form':form})
+
+
 def home(request):
     #side and search bar
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -24,6 +80,8 @@ def room(request,pk):
     return render(request,'base/room.html',  {'room': room})
 
 #create
+@login_required(login_url='login')
+
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -38,10 +96,15 @@ def create_room(request):
     return render(request,'base/room_form.html',context)
 
 #update
+@login_required(login_url='login')
 def update_room(request,pk):
     room = Room.objects.get(id=pk)
-
     form = RoomForm(instance=room)
+
+    if room.host != request.user:
+        return HttpResponse("You are not allowed!!")
+
+
     if request.method =="POST":
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -50,8 +113,14 @@ def update_room(request,pk):
 
     return render(request,'base/room_form.html',  {'form': form})
 
+#delete
+@login_required(login_url='login')
 def delete_room(request,pk):
     room = Room.objects.get(id=pk)
+
+    if room.host != request.user:
+        return HttpResponse("You are not allowed!!")
+    
     if request.method =="POST":
         room.delete()
         return redirect('home')
